// apps/web/src/lib/api/generated/models/aiTypedExtractionValueDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export type AiTypedExtractionValueDto =
	| {
			kind: 'text';
			value: string;
	  }
	| {
			kind: 'number';
			value: number;
	  }
	| {
			kind: 'boolean';
			value: boolean;
	  }
	| {
			kind: 'date';
			value: string;
	  };