// apps/web/src/lib/api/generated/models/extractionFieldDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractionFieldDto {
	field_key: string;
	id: string;
	label: string;
	project_id: string;
	required: boolean;
	value_type: string;
	/** @minimum 0 */
	version: number;
}